rename token switch file, fix isTestnet issue
